+++ src/python4cpm/secrets.py
@@ -3,7 +3,7 @@ from python4cpm.crypto import Crypto
 
 class Secret:
     def __init__(self, secret: str) -> None:
-        self._secret = secret
+        self._value = secret
 
     @classmethod
     def from_env_var(cls, secret: str | None) -> object | None:
@@ -18,5 +18,13 @@ class Secret:
 
     def get(self) -> str:
         if Crypto.ENABLED:
-            return Crypto.decrypt(self._secret)
-        return self._secret
+            return Crypto.decrypt(self._value)
+        return self._value
+
+
+class Password(Secret):
+    pass
+
+
+class NewPassword(Secret):
+    pass